Use chat delete event timestamp

+++ src/storage/sqlstorage/chat_ephemeral_setting.go
@@ -1,6 +1,8 @@
 package sqlstorage
 
 import (
+	"time"
+
 	"github.com/devlikeapro/gows/storage"
 	"go.mau.fi/whatsmeow/types"
 )
@@ -30,10 +32,25 @@ func (s *SqlChatEphemeralSettingStore) UpdateChatEphemeralSetting(setting *stora
 	if setting.IsEphemeral {
 		return s.UpsertOne(setting)
 	} else {
-		return s.DeleteChatEphemeralSetting(setting.ID)
+		return s.DeleteChatEphemeralSetting(setting.ID, time.Now())
 	}
 }
 
-func (s *SqlChatEphemeralSettingStore) DeleteChatEphemeralSetting(id types.JID) error {
-	return s.DeleteById(id.String())
+func (s *SqlChatEphemeralSettingStore) DeleteChatEphemeralSetting(id types.JID, deleteBefore time.Time) error {
+	// First get the current setting to check its timestamp
+	setting, err := s.GetChatEphemeralSetting(id)
+	if err != nil {
+		return err
+	}
+	if setting == nil {
+		return nil
+	}
+	if setting.Setting == nil || setting.Setting.Timestamp == nil {
+		return s.DeleteById(id.String())
+	}
+	// Only delete if the setting is older than the delete event
+	if *setting.Setting.Timestamp < deleteBefore.Unix() {
+		return s.DeleteById(id.String())
+	}
+	return nil
 }
